demo upload

+++ source/PlayState.hx
@@ -9,6 +9,8 @@ import flixel.graphics.FlxGraphic;
 import flixel.group.FlxGroup;
 import flixel.text.FlxText;
 import flixel.tile.FlxTilemap;
+import flixel.tweens.FlxEase;
+import flixel.tweens.FlxTween;
 import flixel.util.FlxColor;
 import hscript.Interp;
 import hscript.Parser;
@@ -27,6 +29,7 @@ class PlayState extends FlxState
 	// hud
 	var levelTM:FlxTilemap;
 	var scoreTxt:FlxText;
+	var debugTxt:FlxText;
 	var coins:FlxGroup;
 
 	function loadstringFile(file:String):String
@@ -37,6 +40,10 @@ class PlayState extends FlxState
 
 	// cam hud
 	public static var camHUD:FlxCamera;
+
+	public static var chapType:String = "tutorial";
+
+	var debugMode:Bool = false;
 
 	override function create()
 	{
@@ -50,15 +57,26 @@ class PlayState extends FlxState
 		FlxG.cameras.add(camHUD, false);
 		
 		levelTM = new FlxTilemap();
-		levelTM.loadMapFromCSV("assets/id/lev" + levelID + "/data.csv",
-			FlxGraphic.fromBitmapData(openfl.Assets.getBitmapData("assets/id/graphic/graphic.png")), 0, 0);
+		levelTM.loadMapFromCSV(Paths.file("id/" + chapType + "/lev" + levelID + ".csv"),
+			FlxGraphic.fromBitmapData(openfl.Assets.getBitmapData("assets/id/" + chapType + "/graphic.png")), 0, 0);
 		add(levelTM);
 
-		player = new FlxSprite(Std.parseFloat(loadstringFile("assets/id/lev" + levelID + "/playerX.txt")),
-			Std.parseFloat(loadstringFile("assets/id/lev" + levelID + "/playerY.txt")));
+		exit = new FlxSprite(35 * 8 + 1, 25 * 8);
+		exit.loadGraphic("assets/images/doors.png", true, 14, 16);
+		exit.animation.add("still", [0]);
+		exit.animation.add("open", [0, 1, 2, 3], 4, false);
+		exit.animation.add("areopen", [3]);
+		trace(exit.x);
+		trace(exit.y);
+		add(exit);
+
+		player = new FlxSprite(/*Std.parseFloat(loadstringFile("assets/id/lev" + levelID + "/playerX.txt")),
+			Std.parseFloat(loadstringFile("assets/id/lev" + levelID + "/playerY.txt")) */);
 		trace(player.x);
 		trace(player.y);
-		player.loadGraphic("assets/images/player.png", false, 8, 8);
+		player.loadGraphic("assets/images/player.png", true, 8, 8);
+		player.animation.add("idle", [0, 1, 2], 1, true);
+		player.animation.play("idle");
 		player.maxVelocity.set(80, 200);
 		player.acceleration.y = 200;
 		player.drag.x = player.maxVelocity.x * 4;
@@ -66,33 +84,31 @@ class PlayState extends FlxState
 		// FlxG.camera.zoom = 1.5;
 		add(player);
 
-		exit = new FlxSprite(35 * 8 + 1, 25 * 8);
-		exit.makeGraphic(14, 16, FlxColor.GREEN);
-		trace(exit.x);
-		trace(exit.y);
-		exit.exists = false;
-		add(exit);
-
 		coins = new FlxGroup();
 		var interp = new Interp();
 		interp.variables.set("createCoin", createCoin);
 		interp.variables.set("setExitPOS", function (x:Int, y:Int) {
-			exit.setPosition(x * 8 + 1, y * 8);
+			exit.setPosition(x, y);
 		});
 		interp.variables.set("setPlayerPOS", function(x:Float, y:Float)
 		{
 			player.x = x;
 			player.y = y;
 		});
+		interp.variables.set("levelID", levelID);
 		interp.variables.set("FlxG", FlxG);
 		var parser = new Parser();
-		interp.execute(parser.parseString(loadstringFile("assets/id/lev" + levelID + "/data.txt")));
+		interp.execute(parser.parseString(loadstringFile("assets/id/" + chapType + "/data.txt")));
 		add(coins);
 
 		scoreTxt = new FlxText(2, 2, 80, "SCORE: 0\nLEVEL: " + levelID);
 		scoreTxt.setFormat(null, 8, FlxColor.WHITE, LEFT, OUTLINE, FlxColor.BLACK);
 		scoreTxt.cameras = [camHUD];
 		add(scoreTxt);
+		debugTxt = new FlxText(scoreTxt.x, scoreTxt.y, 0, "DEBUG MODE");
+		debugTxt.setFormat(null, 8, FlxColor.WHITE, LEFT, OUTLINE, FlxColor.BLACK);
+		debugTxt.cameras = [camHUD];
+		add(debugTxt);
 	}    
 
 	/**
@@ -108,6 +124,36 @@ class PlayState extends FlxState
 	override function update(elapsed:Float)
 	{
 		player.acceleration.x = 0;
+
+		if (FlxG.keys.justPressed.F12) // toggle debug key
+		{
+			debugMode = !debugMode;
+			debugTxt.text = (debugMode ? "DEBUG MODE" : "");
+		}
+
+		if (debugMode)
+		{
+			if (FlxG.mouse.overlaps(exit))
+			{
+				debugTxt.text = "DEBUG MODE\nPOSTION: " + exit.x + " - " + exit.y;
+				if (FlxG.keys.pressed.LEFT)
+				{
+					exit.x -= 1;
+				}
+				if (FlxG.keys.pressed.RIGHT)
+				{
+					exit.x += 1;
+				}
+				if (FlxG.keys.pressed.UP)
+				{
+					exit.y -= 1;
+				}
+				if (FlxG.keys.pressed.DOWN)
+				{
+					exit.y += 1;
+				}
+			}
+		}
 
 		if (FlxG.keys.anyPressed([LEFT, A]))
 		{
@@ -131,7 +177,7 @@ class PlayState extends FlxState
 			openSubState(new PauseSubState());
 		}
 
-		if (FlxG.keys.justPressed.F6)
+		/*if (FlxG.keys.justPressed.F6)
 		{
 			exit.exists = !exit.exists;
 		}
@@ -139,7 +185,7 @@ class PlayState extends FlxState
 		if (FlxG.keys.justPressed.F5)
 		{
 			FlxG.resetState();
-		}
+		}*/
 
 		super.update(elapsed);
 
@@ -150,9 +196,13 @@ class PlayState extends FlxState
 
 	function win(Exit:FlxObject, Player:FlxObject):Void
 	{
-		scoreTxt.text = "SCORE: " + (coins.countDead() * 100) + "\nLEVEL: " + levelID;
-		player.kill();
-		// openSubState(new WinSubState());
+		// check if player is get all coin
+		if (coins.countLiving() == 0)
+		{
+			scoreTxt.text = "SCORE: " + (coins.countDead() * 100) + "\nLEVEL: " + levelID;
+			player.kill();
+			openSubState(new WinSubState());
+		}
 	}
 
 	function getCoin(Coin:FlxObject, Player:FlxObject):Void
@@ -163,7 +213,14 @@ class PlayState extends FlxState
 
 		if (coins.countLiving() == 0)
 		{
-			exit.exists = true;
+			exit.animation.play("open");
+			exit.animation.finishCallback = function(name:String)
+			{
+				if (name == "open")
+				{
+					exit.animation.play("areopen");
+				}
+			}
 		}
 	}
 }
